// src/main/interfaces/responseIf.ts
export interface ackTableDataIF {
  tableId: string;
  minimumSeat: number;
  maximumSeat: number;
  gameType: string;
  seatIndex: number;
  activePlayers: number;
  gameStartTimer: number;
  turnTimer: number;
  tableState: string;
  closedDeck: string[];
  opendDeck: string[];
  turnCount: number;
  dealerPlayer: number;
  declareingPlayer: string;
  validDeclaredPlayer: string;
  validDeclaredPlayerSI: number;
  playersDetail: userSeatKeyIf[];
  playersCount: number;
}

interface signupResponseIF {
  _id: string;
  un: string;
  pp: string;
  socketid: string;
  tableId: string;
  gameId: string;
  lobbyId: string;
  isFTUE: boolean;
  // chips: number;
  isPlay: boolean;
  fromBack: boolean;
  isRobot: boolean;
  dealType: number;
  gamePoolType: number;
  latitude: string;
  longitude: string;
  entryFee: number
  maximumSeat: number,
  maxTableCreateLimit: number,
  rummyType : string;
  authToken: string;
  isLink?:boolean;
  isCreateRoom : boolean;
  isReferralCode : string;
  isPrivateTableContinue? : boolean;
  gameEntryFee?:number;
  isCash:boolean;
  coins:number;
  cash:number;
  winCash:number;
  bonus:number;
  avatarName:string;
}

export interface formatSignupAckIf {
  signupResponse: signupResponseIF;
  gameTableInfoData: ackTableDataIF[];
  reconnect: boolean;
}